# Weights for 1.2.5 - `coretime-kusama` + `kusama` fix for new prefixed with `polkadot_runtime_common`